Create issue tracker to assign days/notes to create
This is a simple issue tracker or assignment page that keeps track
of who has selected what days notes to cover. The "Present Day' tag
is simply there to keep track of what day has currently passed. To
take a day simply follow the header in organization and type your
name to the day you wish to cover. There are 24 days and 6 of us
so each person takes 4 days. At the current date not all days are
available so you must not take a number of day to where someone
cannot do the check in.