First Draft of Day 2 Notes
This is the first version of the Day 2 Notes. This class was mostly
focused on command line tools and getting comfortable working in
the shell.

The notes cover an introduction to the shell, how to connect to the
remote servers, a list of commonly used commands, how to use basic
for loops in the shell, and some commonly used utilities such as
globbing and piping.

fix my name and email in the git config settings