Detail lecture 4 notes
I have created the lecture 4 notes using an .md (markdown file). The
file is split into three sections: commands, programs, and notes. The
command section is for any commands learned during that particular
lecture, the programs are any programs learned during this particular
lecture and the note section are the general notes made from the
lecture.