Add Lecture 6 notes
This is the first draft of lecture 6 notes. This lecture was the first
major introduction to version control systems.

We learned about the difference between centralized and distributed
systems, as well as their pros and cons. There was more focus on DVCS,
ahow this architecture keeps track of file changes to help us delve
deeper into git.