Add lecture 6 notes
This is the first draft for lecture 6's notes. In this class, we
discussed version control systems.

We first discussed goals achieved through VCSs, the centralized and
distributed architectures, and how DVCSs keep track of file changes.
Also, Prof. Angstadt quickly demonstrated how to add, commit, and push
files to a repo, specifically on the linux servers.