Add lecture 6 notes
This is the first draft for lecture 6's notes. In this class, we
discussed version control systems.

We first talked about the goals achieved through using VCSs and the
differences between the centralized and distributed architectures. We
proceeded to discuss how DVCSs keep track of file changes through
snapshots. Lastly, Prof. Angstadt quickly demonstrated how to add,
commit, and push files to a git repo.