Update Issue Tracker
This will add a markdown file that will easily show who is currently
working on what day/lecture. It is in a neat orderly table. To first
take a day/issue please reference the issues first and claim a day
there. Then update this file when you do your notes. Please remember to
commit this file alongside the notes being pushed.