Add lecture 18
This is the first draft of lecture 18's notes. In this class, we
discussed plans to make our fault localizer file more efficient.

To do this, Prof. Angstadt discussed about the differences between the
Tarantula and Ochiai metrics, implementing these methods to our
project. He also briefly demonstrated how to conduct a binary search (or
bisect) to find bugs within a repo.