Add the Template file
Add markdown file to be used as a template for review and all
future notes. Open for edits until a date specified in the
related issue. After the file is final, all notes can be
edited to conform.

Template file describes convention for using example code as
both python and bash code, details an organization based on
main topics and a general commands at the end.