Hyperlink P3 in lecture 14's notes
Create hyperlink for project 3.

I've written everything to be as synchronous with the class as possible.
In this updated draft, I've decided to add the last line to mention that
the end of this lecture is completely devoted to working on P3, with a
hyperlink to the project from Sakai.