Adjust title and bold headers lecture 18's notes
Change the title format and lecture 18's note, as well as fixing errors
with bolding in two locations.

In this class, we discussed debugging as hypothesis. Topics include
implementing the Tarantula and Ochiai metrics into our code, what it
takes to create an efficient bug report, and looking into conducting
bisecting searches in git.

Similar to my other recommits, I'll readjust the title for each
lecture. I've also reformatted the line beginning with "What is the
math...", as well as breaking the line beginning with "One can
conduct..." to make things seem less conspicuous.