Reformat title of  lecture 1's notes
Change the title for lecture 1's notes.

In this class, we were introduced to software engineering. We discussed
how critical software maintenance is to our daily lives, using "software
disasters" to exemplify how software can affect us on large scales. We
also discussed the differences between software and a program, as well
as managing sotware on a high-level.

In this commit, I've reformatted the title of the notes to conform to
what Olivia has recommended. I believe it looks neater and streamlined.