Re-title lecture 14's notes
Adjusting the title of lecture 14's notes.

In this class, we wrapped up our discussion on quality assurance. The
largest topic of this lecture was determining how we measure quality.
The remainder of the class was dedidcated to P3, which I've hyperlinked
in the last commit.

Like the rest of my other requests, I've reformatted the title for the
notes, indicating which class they're from, the main topic, and the
date. Everything else remains the same.