Add Git Review and Initial Draft of Git Branching to Lecture 07
Lecture_07.md header was not updated with the appropriate date and topic
information so those were added accordingly. The Git Review section was
completed and contains content related to commits and the Git tree. The
Git Branching section was also added with content up to merging in
timestamp <1:12:47> in the Panopto CS-340 Su21 Lecture 7 Video.