Update and Modify Lecture_19.md
Delete the 'Overview' section from the notes and specify the language
of code for each code block in the Lecture_19.md file.

'Overview' no longer conforms to the current formatting because it is
not needed, and was addressed in issue #33. Specifying the code
language is in response to changes suggested for template.md.