Update Day2.md name and formatting
Change name of Day2.md to Lecture_02.md to fit current name convention
as detailed in issue #9. Add header 'CS 340 summer notes', as well as
the date and unify topic headers to all be the same size. Also, declare
language of code for code blocks, as specified in the current formatting.