Update Title Format and Header Spacing
Fix spacing between header and rest of notes to be a single line. Update
topic titles from four hashes (####) to two (##). This allows for
greater granularity of topic breakdown via the use of triple or
quadruple hashes for subsections. This format is also more consistent
with previous commits such as 66c1f81 for Day2.md, f5ff403 from
Lecture_01.md, and 48f0171 from Lecture_04.md.

Add a bullet point with an indented subsection to provide an support for
these as well.