Add Input Generation and Finish Lecture 12 Notes
Finish the Mutation Testing section by adding a subsection on the
limitations of the method. Add a section on Input Generation with
subsections for test input and comparators. Add an example python code
block to explain branch and path coverage, this will serve as motivation
for theorem provers and automating test input. Remove remaining code
from the template.