Add Review Section to Lecture 20
Lecture 20 starts in the middle of the delta debugging conversation with
an exercise and review of our pseudo-implementation of delta debugging.
Add a Delta Debugging (DD) Review section highlighting some of the main
insights of this implementation. Add code block with the pseudo-python
implementation of DD from class before, this was used in the exercise
and referenced in the beginning.