Add Bonus: CS Tangents Feature to Lecture 20
Lecture 20 concludes with tangents on the Theory of Computation and
Virtual memory. These are not explicitly relevant to the course syllabus
but can still aid in general CS knowledge. Bundle these points and add a
Bonus: Other CS tangents section to Lecture_20.md.

Note that the formatting does not have subsections titles that may
otherwise seem intuitive to have, namely for Theory of Computation and
OS. These have a bullet point hierarchy instead, which is intended to
emphasize how this is bonus content and to not draw much attention with
big titles.

Remove lasting template content.