Reorganize Sub-Sections in Input Gen for Lecture 12
The previous version of Lecture_12.md had the subsections in test input
generation follow the structure of the definition, namely Test Input
before Comparator. However, the chronological ordering of these sections
(which is the reverse) is more intuitive and portrays a more accurate
picture of the logic followed in class.