Re-Format The Mutants Section in Lecture 12
Lecture_12.md contains a section on Mutation Testing where the Competent
Programmer Hypothesis (CPH) was an important point (hence in bold)
within the Mutants subsection. Update this to make the CPH its own
sub-section (triple hash ###). CPT is an important topic that also
contains significant content both explaining it and related to it.

Update from bullets to paragraphs. Bullet point content was written in
sentences, and the previous answer format was more fitting to bullets
than the section format, where prose fairs better.