Update Review Column and Add Reviewer Names
The 'Reviewed' column was unclear as to whether it referred to PR
reviews or to editing the format of specific files. Since the PR
reviews are more formally assigned, that is what was chosen for
this file to track. The names for each completed PR review that ended
in a merge have been added, and the column itself had its name changed
to 'PR Review' so it is more clear.