Add date and fix bullet point formatting
The date was added to the Lecture_08 file from the date listed on
Sakai. Also, there was a problem with the bullet points in the file,
so I added a line to make the formatting work. This is the same as
an earlie commit, but on a new branch so that it has no connection to
other unmerged files.