Update Resources to Citations in README
The previous README had a section on resources for the repo. While
various resources were used during the scope of project 2, the contents
of the repository themselves are mostly taken from class lectures.
Therefore update the resources to citations to reflect how this space is
better used to attribute content to the course itself (and other
resources if still important). Other resources are better noted in
individual project submissions.

Note that the format for the citations was taken from MIT's citation
guideline for their open courses online:
https://ocw.mit.edu/help/faq-cite-ocw-content/. Reformat with
information from the course website.