Add Media Folder and Prof Logo
Add a Media directory to the repository and a png logo extracted from
the instructor's class website. These will be used to add the Angstadt
logo to the header of the README file, to link visually to the class and
the instructor's website. This is also consistent with other README
files.

While not used in this specific iteration of the course, a media
directory might be a good inspiration to have a relative path for photos
in lecture notes, eg. diagrams, graphs, etc. This directory could be
shown as an example of this, or for later additions if this repository
spans after the scope of Project 2.